Update 2019_12_14_pkt-whitepaper-website.md
I am asking that this submission be granted as Milestone 1 & 3 combined. This is because an initial whitepaper was submitted and rejected. So technically a review process has already occurred. If there are any comments that need to be addressed to make this whitepaper correct, please let me know. If you think this whitepaper is acceptable, I would like for this to be accepted as Milestone 1& 3.