Update 2022_03_04_pkt_technical_documentation.md
I am submitting the project milestone for technical documentation for PKT. When creating this proposal I expected it to take 133 hours to complete which it did. I also added a few extra steps for setting up a CLI wallet as I felt this was important to include so that miners could easily be able to have all the information when setting up a CLI Wallet. Please let me know if there is anything else I need to include. 

Thank you.