Update on PktDesk and Request for 350,000 PKT
I wanted to update you on the progress of PktDesk, and to request the release of 350,000 PKT as promised.

Over the past year, PktDesk has helped over 4500+ people to learn about PKT, according to Google Console. We have added 11 new articles and updated the old ones to ensure that the content is informative and up-to-date. As a result, the site is still enjoying 1st position for the majority of the keywords in Google.

If there is anything missing from the website or if there are any changes you would like to see, please do let me know. I would be happy to make any necessary changes to ensure that the website meets your expectations.

We are committed to continuing to improve PktDesk, and to reach an even wider audience. To help us do that, we kindly request that you release the promised 350,000 PKT. This payment will go a long way in helping us to cover our expenses and invest in further development.

Thank you for your continued support, and we look forward to hearing from you soon.